1.连接判断

+++ spring-support-monitor-server-starter/src/main/java/com/chua/starter/monitor/server/controller/gen/SessionController.java
@@ -226,6 +226,10 @@ public class SessionController {
         StringBuilder stringBuffer = new StringBuilder();
         long startTime = System.nanoTime();
         Session session = ServiceProvider.of(Session.class).getKeepExtension(executeQuery.getGenId(), sysGen.getGenType(), sysGen.newDatabaseOptions());
+        if(!session.isConnect()) {
+            ServiceProvider.of(Session.class).closeKeepExtension(executeQuery.getGenId() );
+            return ReturnResult.illegal("当前服务器不可达");
+        }
         SessionResultSet sessionResultSet = null;
         Map<String, String> remark = new HashMap<>();
         try {
@@ -327,6 +331,10 @@ public class SessionController {
         }
 
         Session session = ServiceProvider.of(Session.class).getKeepExtension(sysGen.getGenId() + "", sysGen.getGenType(), sysGen.newDatabaseOptions());
+        if(!session.isConnect()) {
+            ServiceProvider.of(Session.class).closeKeepExtension(sysGen.getGenId() + "");
+            return ReturnResult.illegal("当前服务器不可达");
+        }
         SessionInfo sessionInfo = null;
         try {
             sessionInfo = session.info();
@@ -367,6 +375,10 @@ public class SessionController {
         }
 
         Session session = ServiceProvider.of(Session.class).getKeepExtension(saveQuery.getGenId(), sysGen.getGenType(), sysGen.newDatabaseOptions());
+        if(!session.isConnect()) {
+            ServiceProvider.of(Session.class).closeKeepExtension(saveQuery.getGenId());
+            return ReturnResult.illegal("当前服务器不可达");
+        }
         SessionInfo sessionInfo = null;
         File file1 = null;
         try {
@@ -407,6 +419,10 @@ public class SessionController {
         }
 
         Session session = ServiceProvider.of(Session.class).getKeepExtension(deleteQuery.getGenId(), sysGen.getGenType(), sysGen.newDatabaseOptions());
+        if(!session.isConnect()) {
+            ServiceProvider.of(Session.class).closeKeepExtension(deleteQuery.getGenId());
+            return ReturnResult.illegal("当前服务器不可达");
+        }
         SessionInfo sessionInfo = null;
         try {
             sessionInfo = session.delete(deleteQuery);
@@ -450,6 +466,10 @@ public class SessionController {
         } catch (Exception ignored) {
         }
         Session session = ServiceProvider.of(Session.class).getKeepExtension(updateQuery.getGenId(), sysGen.getGenType(), sysGen.newDatabaseOptions());
+        if(!session.isConnect()) {
+            ServiceProvider.of(Session.class).closeKeepExtension(updateQuery.getGenId());
+            return ReturnResult.illegal("当前服务器不可达");
+        }
         SessionInfo sessionInfo = null;
         try {
             sessionInfo = session.update(updateQuery, file1);
